+++ mod_form.php
@@ -55,11 +55,7 @@ class mod_mumie_mod_form extends moodleform_mod {
         $mform = &$this->_form;
 
         $this->servers = $this->get_valid_servers_with_structure();
-        $serveroptions = array();
-        $courseoptions = array();
-        $languageoptions = array();
-
-        self::populate_options($serveroptions, $courseoptions, $languageoptions);
+        $serveroptions = $this->get_server_options();
 
         // Adding the "general" fieldset, where all the common settings are shown.
         $mform->addElement('header', 'mumie_multi_edit', get_string('mumie_form_activity_header', 'mod_mumie'));
@@ -89,7 +85,13 @@ class mod_mumie_mod_form extends moodleform_mod {
             );
         }
 
-        $mform->addElement("select", "mumie_course", get_string('mumie_form_activity_course', "mod_mumie"), $courseoptions);
+        $mform->addElement(
+            "text",
+            "mumie_course",
+            get_string('mumie_form_activity_course', "mod_mumie"),
+            array("disabled" => true, "class" => "mumie_text_input")
+        );
+        $mform->setType("mumie_course", PARAM_TEXT);
 
         $mform->addElement("hidden", "language", $USER->lang, array("id" => "id_language"));
         $mform->setType("language", PARAM_TEXT);
@@ -248,62 +250,16 @@ class mod_mumie_mod_form extends moodleform_mod {
     }
 
     /**
-     * For the form to work, we need to populate all option fields before javascript can modify them.
-     * If a option is not defined before js is started, it won't be saved by moodle.
+     * Get all options for server drop-down menu
      *
-     * @param stdClass $serveroptions pointer to the array containing all available servers
-     * @param stdClass $courseoptions pointer to the array containing all available courses
-     * @param stdClass $languageoptions pointer to the array containing all available languages
-     * @return void
+     * @return array
      */
-    private function populate_options(&$serveroptions, &$courseoptions, &$languageoptions) {
+    private function get_server_options() {
+        $serveroptions = array();
         foreach ($this->servers as $server) {
             $serveroptions[$server->get_urlprefix()] = $server->get_name();
-            self::populate_course_options(
-                $server->get_courses(),
-                $courseoptions,
-                $languageoptions
-            );
         }
-    }
-
-    /**
-     * Populate course option list and then populate task options for all given courses of a MUMIE server.
-     *
-     * @param array $courses array containing a list of courses
-     * @param array $courseoptions pointer to the array containing all available courses
-     * @param array $languageoptions pointer to the array containing all available languages
-     * @return void
-     */
-    private function populate_course_options($courses, &$courseoptions, &$languageoptions) {
-        foreach ($courses as $course) {
-            foreach ($course->get_name() as $name) {
-                $courseoptions[$name->value] = $name->value;
-
-                // If a user wants to use an entire course instead of a single problem, we need to define a pseudo problem to use.
-                $languagelink = $course->get_link() . '?lang=' . $name->language;
-                $languageoptions[$name->language] = $name->language;
-            }
-            self::populate_problem_options($course, $languageoptions);
-        }
-    }
-
-    /**
-     * Populate task and language option list for a given course
-     *
-     * @param \auth_mumie\mumie_course $course single instance of MUMIE course containing a list of tasks
-     * @param array $languageoptions pointer to the array containing all available languages
-     * @return void
-     */
-    private function populate_problem_options($course, &$languageoptions) {
-        foreach ($course->get_tasks() as $problem) {
-            $link = $problem->get_link();
-
-            foreach ($problem->get_headline() as $headline) {
-                $languagelink = $link . '?lang=' . $headline->language;
-                $languageoptions[$headline->language] = $headline->language;
-            }
-        }
+        return $serveroptions;
     }
 
     /**